requirements.txt
google-auth>=2.0.0
google-api-python-client>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
